+++ README_old.md
@@ -1,0 +1,31 @@
+# KnowNet
+
+A project building for knowledge representation in scientific articles
+
+Build status: [![Build Status](https://travis-ci.com/PKU-Dragon-Team/Project-KnowNet.svg?branch=master)](https://travis-ci.com/PKU-Dragon-Team/Project-KnowNet)
+
+## Current Status
+
+### data_fetcher
+
+To be finished.
+
+### network_construction
+
+**Working** (details to be finished)
+
+### network_analysis
+
+**Working** (details to be finished)
+
+### data_platform
+
+1. document: **Working**
+2. datasource: _Partially Working_
+    - json (DocDataSource): **Working** (except for _wildcard filtering_)
+    - networkx (GraphDataSource): **Working** (except for _wildcard filtering_)
+    - sqlite (RowDataSource): **Working** (except for _wildcard filtering_)
+    - science_direct (DocDataSource): **Working** (except for _wildcard filtering_)
+3. utility: Not yet implemented
+4. config: **Working**
+5. log: Not yet implemented
