+++ network_construction/README_old.md
@@ -1,0 +1,9 @@
+请参照数据处理模块B的word文档进行理解各个模块，有的模块的参数与文档有些许差异，以代码实现为准。
+
+1. algorithm模块独立出来，提供了一些文本挖掘的方法。
+2. database模块专门用于封装数据库操作，所有的查询、读取、写入等都在这里封装，别的模块的数据库操作全都使用此模块即可。
+3. source模块用于获取数据源，从数据源中得到文章的作者、引文、全文等信息，并将这些信息提供给别的模块。
+4. node模块专门用于抽取各类节点
+5. relation模块用于抽取各类关系
+6. network模块用于调用node和relation模块，直接生成网络，属于本部分代码中最上层的部分。
+7. 想要构建网络，只需要import network，调用其中的构建网络函数即可。但是这之前需要先建立一个图数据库（使用database模块中的方法）。
